requirements.txt
#api
fastapi[standard]
uvicorn[standard]
asyncpg
python-multipart
fastapi-mail

#database and datap pipline 
SQLAlchemy 
aiosqlite
alembic
redis
#auth
python-jose[cryptography]
bcrypt 

#simulation
requests